+++ src/plugins/blocksmith/components/instagram-media-block-modal/instagramMediaBlockRegex.js
@@ -1,1 +1,1 @@
-export default /https?:\/\/www\.instagram\.com\/(p|tv)\/[a-zA-z0-9]+/;
+export default /https?:\/\/www\.instagram\.com\/(p|tv)\/[a-zA-z0-9-]+/;