+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 * Correct `pbjUrl` template for poll and widget pickers.
 * Fix structure > teasers if pressing 1 to change order date, the selector/cursor will lose focus.
 * Add `autocomplete: true` to the search nodes request in node picker field.
+* Issue #1094 Instagram block not allowing id's with a dash
 
 
 ## v0.2.8