+++ CHANGELOG.md
@@ -1,8 +1,9 @@
 # CHANGELOG
 
 
-## patch
+## vX.X.X
 * Can edit ad-widget size while in view mode
+* Remove error message when editing/creating slugs
 
 
 ## v0.2.10