+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 
 ## vX.X.X
+* Fixed instagram media block not allowing hyphens in embed url.
 * Can edit ad-widget size while in view mode
 * Remove error message when editing/creating slugs
 * Re-enable collaboration reducer on `MESSAGE_RECEIVED`.